refactored config to separate go module
